+++ handlers/nonce/checker/nonce.go
@@ -1,20 +1,52 @@
 package noncechecker
 
 import (
+	"fmt"
 	"strconv"
+	"strings"
 
 	log "github.com/sirupsen/logrus"
 	"gitlab.com/ConsenSys/client/fr/core-stack/orchestrate.git/ethereum/ethclient"
 	"gitlab.com/ConsenSys/client/fr/core-stack/orchestrate.git/pkg/engine"
 	"gitlab.com/ConsenSys/client/fr/core-stack/orchestrate.git/pkg/errors"
 	"gitlab.com/ConsenSys/client/fr/core-stack/orchestrate.git/services/nonce"
+	ierror "gitlab.com/ConsenSys/client/fr/core-stack/orchestrate.git/types/error"
 	"gitlab.com/ConsenSys/client/fr/core-stack/orchestrate.git/types/ethereum"
 )
 
+func controlRecoveryCount(txctx *engine.TxContext, conf *Configuration) error {
+	var count int
+	v, ok := txctx.Envelope.GetMetadataValue("nonce.recovering.count")
+	if ok {
+		i, err := strconv.Atoi(v)
+		if err != nil {
+			return err
+		}
+		count = i
+	}
+
+	if count >= conf.MaxRecovery {
+		// If maximum recovery count is reached do not recover
+		return fmt.Errorf("nonce: reached max recovery count")
+	}
+
+	// Incremenent recovery count on envelope
+	txctx.Envelope.SetMetadataValue("nonce.recovering.count", fmt.Sprintf("%v", count+1))
+	return nil
+}
+
 // Checker creates an handler responsible to check transaction nonce value
 // It makes sure that transactions with invalid nonce are not processed
-func Checker(nm nonce.Sender, ec ethclient.ChainStateReader) engine.HandlerFunc {
+func Checker(conf *Configuration, nm nonce.Sender, ec ethclient.ChainStateReader, tracker *RecoveryTracker) engine.HandlerFunc {
 	return func(txctx *engine.TxContext) {
+		if mode, ok := txctx.Envelope.GetMetadataValue("tx.mode"); ok && mode == "raw" {
+			// If transaction has been generated externally we skip nonce check
+			txctx.Logger.WithFields(log.Fields{
+				"metadata.id": txctx.Envelope.GetMetadata().GetId(),
+			}).Debugf("nonce: skip check for raw transaction")
+			return
+		}
+
 		// Retrieve chainID and sender address
 		chainID, sender := txctx.Envelope.GetChain().ID(), txctx.Envelope.GetFrom().Address()
 		txctx.Logger = txctx.Logger.WithFields(log.Fields{
@@ -31,16 +63,15 @@ func Checker(nm nonce.Sender, ec ethclient.ChainStateReader) engine.HandlerFunc 
 		lastSent, ok, err := nm.GetLastSent(nonceKey)
 		if err != nil {
 			e := txctx.AbortWithError(err).ExtendComponent(component)
-			txctx.Logger.WithError(e).Errorf("nonce: could not retrieve last sent nonce")
+			txctx.Logger.WithError(e).Errorf("nonce: could not load last sent nonce")
 			return
 		}
 
-		// If no nonce is available (meaning that envelope being processed is the first one for a given sender on a given chain)
-		// then we calibrate nonce manager by retrieving nonce from chain
 		if ok {
 			expectedNonce = lastSent + 1
 		} else {
-			// Retrieve nonce from chain
+			// If no nonce is available (meaning that envelope being processed is the first one for the pair sender,chain)
+			// then we retrieve nonce from chain
 			pendingNonce, err := ec.PendingNonceAt(txctx.Context(), chainID, sender)
 			if err != nil {
 				e := txctx.AbortWithError(err).ExtendComponent(component)
@@ -49,17 +80,14 @@ func Checker(nm nonce.Sender, ec ethclient.ChainStateReader) engine.HandlerFunc 
 			}
 			txctx.Logger.WithFields(log.Fields{
 				"nonce.pending": pendingNonce,
-			}).Debugf("nonce: retrieving nonce from chain")
+			}).Debugf("nonce: calibrating nonce from chain")
 
 			expectedNonce = pendingNonce
 		}
 
-		// Compare expected nonce to attributed nonce
-		// to make sure we do not sent a transaction with invalid nonce
 		n := txctx.Envelope.GetTx().GetTxData().GetNonce()
 		if n != expectedNonce {
-
-			// Transaction has an invalid nonce
+			// Attributes nonce is invalid
 			txctx.Logger.WithFields(log.Fields{
 				"nonce.expected": expectedNonce,
 				"nonce.got":      n,
@@ -68,24 +96,23 @@ func Checker(nm nonce.Sender, ec ethclient.ChainStateReader) engine.HandlerFunc 
 			// Reset tx nonce, hash and raw
 			resetTx(txctx.Envelope.GetTx())
 
-			if n > expectedNonce {
-				// Nonce is too high
+			// Control recovery count
+			err := controlRecoveryCount(txctx, conf)
+			if err != nil {
+				e := txctx.AbortWithError(err).ExtendComponent(component)
+				txctx.Logger.WithError(e).Errorf("nonce: abort transaction execution")
+				return
+			}
 
-				// Retrieve recovery status
-				recovering, err := nm.IsRecovering(nonceKey)
-				if err != nil {
-					e := txctx.AbortWithError(err).ExtendComponent(component)
-					txctx.Logger.WithError(e).Errorf("nonce: could not load recovery status")
-					return
-				}
-				if !recovering {
-					txctx.Logger.WithFields(log.Fields{
-						"nonce.recovering.expected": expectedNonce,
-					}).Warnf("nonce: start recovering")
-					// Envelope being processed is the first one we encounter with Nonce too High
-					// Indicate expected nonce in metadata to signal tx-nonce worker from which value re-start attributing nonces
+			if n > expectedNonce {
+				// Nonce is too-high
+				if _, ok := txctx.Envelope.GetMetadataValue("nonce.recovering.expected"); ok || tracker.Recovering(nonceKey) == 0 {
+					// Envelope has already been recovered or it is the first one to be recovered
 					txctx.Envelope.SetMetadataValue("nonce.recovering.expected", strconv.FormatUint(expectedNonce, 10))
 				}
+			} else {
+				// If nonce is to low we remove any recovery signal in metadata (possibly coming from a prior execution)
+				delete(txctx.Envelope.GetMetadata().GetExtra(), "nonce.recovering.expected")
 			}
 
 			// We set a context value to indicate to other handlers that
@@ -98,11 +125,18 @@ func Checker(nm nonce.Sender, ec ethclient.ChainStateReader) engine.HandlerFunc 
 			return
 		}
 
+		// If nonce was valid we remove recovery signal in metadata (possibly coming from a prior execution)
+		delete(txctx.Envelope.GetMetadata().GetExtra(), "nonce.recovering.expected")
+		delete(txctx.Envelope.GetMetadata().GetExtra(), "nonce.recovering.count")
+
 		// Execute pending handlers
 		txctx.Next()
 
 		// If pending handlers executed correctly we increment nonce
 		if len(txctx.Envelope.GetErrors()) == 0 {
+			// Possibly re-initiliaze recovery counter
+			tracker.Recovered(nonceKey)
+
 			// Increment last sent nonce
 			err := nm.SetLastSent(nonceKey, n)
 			if err != nil {
@@ -113,9 +147,61 @@ func Checker(nm nonce.Sender, ec ethclient.ChainStateReader) engine.HandlerFunc 
 				// meaning that transaction has been sent to the chain, so we log
 				// error and proceed to next envelope
 				e := errors.FromError(err).ExtendComponent(component)
-				txctx.Logger.WithError(e).Errorf("nonce: could not set nonce on cache")
+				txctx.Logger.WithError(e).Errorf("nonce: could not store last sent nonce")
 			}
+
+			return
 		}
+
+		var errs []*ierror.Error
+		for _, err := range txctx.Envelope.GetErrors() {
+			// TODO: update EthClient to process and standardize nonce too low errors
+			if !strings.Contains(err.String(), "nonce too low") {
+				errs = append(errs, err)
+				continue
+			}
+
+			// We got a "nonce too low" error when sending the transaction
+			txctx.Logger.WithFields(log.Fields{
+				"nonce.got": n,
+			}).Warnf("nonce: invalid nonce")
+
+			// Control recovery count
+			err := controlRecoveryCount(txctx, conf)
+			if err != nil {
+				e := txctx.AbortWithError(err).ExtendComponent(component)
+				txctx.Logger.WithError(e).Errorf("nonce: abort transaction execution")
+				return
+			}
+
+			// We recalibrate nonce from chain
+			pendingNonce, err := ec.PendingNonceAt(txctx.Context(), chainID, sender)
+			if err != nil {
+				e := txctx.AbortWithError(err).ExtendComponent(component)
+				txctx.Logger.WithError(e).Errorf("nonce: could not read nonce from chain")
+				return
+			}
+			txctx.Logger.WithFields(log.Fields{
+				"nonce.pending": pendingNonce,
+			}).Debugf("nonce: calibrating nonce from chain")
+
+			// Re-calibrate cache
+			err = nm.SetLastSent(nonceKey, pendingNonce-1)
+			if err != nil {
+				e := errors.FromError(err).ExtendComponent(component)
+				txctx.Logger.WithError(e).Errorf("nonce: could not store last sent nonce")
+			}
+
+			// Reset tx nonce, hash and raw
+			resetTx(txctx.Envelope.GetTx())
+
+			// We set a context value to indicate to other handlers that
+			// an invalid nonce has been processed
+			txctx.Set("invalid.nonce", true)
+		}
+
+		// Update Envelope errors with Nonce too Low error removed
+		txctx.Envelope.Errors = errs
 	}
 }
 
@@ -127,46 +213,23 @@ func Checker(nm nonce.Sender, ec ethclient.ChainStateReader) engine.HandlerFunc 
 //
 // Setting recovery status before producing the envelope could result in case of crash
 // in a situation were would never be able to signal tx-nonce to recalibrate nonce
-func RecoveryStatusSetter(nm nonce.Sender) engine.HandlerFunc {
+func RecoveryStatusSetter(nm nonce.Sender, tracker *RecoveryTracker) engine.HandlerFunc {
 	return func(txctx *engine.TxContext) {
 		// Execute pending handlers
 		txctx.Next()
 
-		// Retrieves nonce key for nonce manager processing
+		// Compute nonce key
 		nonceKey := string(txctx.In.Key())
 
-		// Test if we have entered recovery mode
-		// In which case we set recovery status to true
+		// If are recovering we increment count
 		if _, ok := txctx.Envelope.GetMetadataValue("nonce.recovering.expected"); ok {
-			// We encountered a first nonce too high scenario so we set recovery status to true
-			err := nm.SetRecovering(nonceKey, true)
-			if err != nil {
-				txctx.Logger.WithError(
-					errors.FromError(err).ExtendComponent(component),
-				).Errorf("nonce: could not load recovery status")
-				return
-			}
+			tracker.Recover(nonceKey)
 		}
 
 		if b, ok := txctx.Get("invalid.nonce").(bool); len(txctx.Envelope.GetErrors()) == 0 && (!ok || !b) {
 			// Transaction has been processed properly
 			// Deactivate recovery if activated
-			recovering, err := nm.IsRecovering(nonceKey)
-			if err != nil {
-				txctx.Logger.WithError(
-					errors.FromError(err).ExtendComponent(component),
-				).Errorf("nonce: could not load recovery status from cache")
-			}
-			if recovering {
-				// Envelope being processed is valid meaning nonce recovery
-				// has completed. So we deactivate recovery status
-				err := nm.SetRecovering(nonceKey, false)
-				if err != nil {
-					txctx.Logger.WithError(
-						errors.FromError(err).ExtendComponent(component),
-					).Errorf("nonce: could not set recovery status from cache")
-				}
-			}
+			tracker.Recovered(nonceKey)
 		}
 	}
 }